services/auth-service/requirements.txt
fastapi
python-dotenv
authlib
httpx
pydantic[email]
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib
itsdangerous
